feat(board): crete mouse dedicated utils and increase utils features
